syntax: stop paying padding and a 64-bit offset on every word (#2073)
A warm interactive startup keeps ~30MB of retained tree, and the profile
says the remaining time is `madvise` and `memclr` — the cost of growing
and zeroing a heap sized by what is *kept*, not by what is thrown away.
Two measurements on this issue already showed that cutting garbage buys
nothing; this cuts what is kept.

Two changes, neither of which moves a semantic:

- `Span`'s fields were declared in reading order — a kind, a value, a
  quoting — and Go padded that out to 88 bytes with 15 bytes of holes.
  Wide fields first and the one-byte ones together at the end makes it
  72. There are ~185,000 spans in the tree a real startup leaves, so the
  holes alone were 3MB of resident memory holding nothing.

- `Pos` was three `int`s, 24 bytes, and it is the most repeated shape in
  a parsed program: a `Word` holds two and every `Span` inside it holds
  another. At `int32` a Pos is 12 bytes, a Word is 48 rather than 72 and
  a Span is 64 rather than 72 — and 48 and 64 are Go size classes, where
  72 was three quarters of one and wasted the rest.

Measured over the 6.3MB of shell the startup path actually reads — 589
files under ~/.zi and ~/.zshrc, parsed and retained:

                   allocated     live at rest
    before           139.2MB          71.9MB
    after            110.5MB          57.5MB
                     -20.6%          -20.1%

The bound `int32` buys that with is 2GiB of input per file, and the
parser already holds the whole of its input as one string, so that is
not the limit anything reaches first.

+++ interp/runner.go
@@ -1987,7 +1987,7 @@ func (r *Runner) lineOf(p syntax.Pos) int {
 		// so the node's own line says nothing.
 		return r.linePin
 	}
-	return p.Line + r.lineBase
+	return int(p.Line) + r.lineBase
 }
 
 // builtinIsSpeaking reports whether this diagnostic belongs to a builtin,
@@ -2177,7 +2177,7 @@ func (r *Runner) RunPart(ctx context.Context, f *syntax.File) error {
 	if r.started.IsZero() {
 		r.started = r.Now()
 	}
-	r.programEnd = f.End().Line + 1
+	r.programEnd = int(f.End().Line + 1)
 	abandoned := 0
 	for _, st := range f.Stmts {
 		if abandoned != 0 && r.lineOf(st.Pos()) == abandoned {
@@ -2464,7 +2464,7 @@ func shiftParseError(err error, by int) error {
 		return err
 	}
 	moved := *se
-	moved.Pos.Line += by
+	moved.Pos.Line += int32(by)
 	if moved.ConstructLine > 0 {
 		moved.ConstructLine += by
 	}
